Fix warnings in cxx_api_tests and llvm_testing.  Also, add in header files so they can be searched by IDE.

+++ source/testing/src/rrTestUtils.cpp
@@ -17,7 +17,7 @@ DoubleMatrix ParseMatrixFromText(const string& textMatrix)
         {
         	if(!mat.size())
             {
-                mat.resize(rows.size(), values.size());
+                mat.resize(static_cast<unsigned int>(rows.size()), static_cast<unsigned int>(values.size()));
             }
 
             mat(row, col) = toDouble(values[col]);
